Refactor UI structure and update import paths
- Moved UI-related modules into a new `ui` directory for better organization, including `notification_popup.py`, `startup_dialogs.py`, and `ui_helpers.py`.
- Updated import paths across various modules to reflect the new structure, ensuring consistent access to UI functionalities.
- Enhanced the Python import helper with updated paths for improved code completion related to UI components.
- Adjusted the main application initialization to accommodate the new UI structure.
