+++ utils/__init__.py
@@ -1,6 +1,6 @@
 # utils_folder/__init__.py
 
 from .settings import Settings
-from . import notification_popup
+from .ui import notification_popup
 
 __all__ = ["UR10_Server_functions", "global_vars", "Settings"]
