+++ utils/app_initialization.py
@@ -1,4 +1,3 @@
-
 """
 Application initialization functions for the Multipack Parser app.
 """
@@ -16,7 +15,7 @@ from PySide6.QtGui import QPixmap, QPainter, QColor
 from utils import global_vars
 from utils.app_control import show_instant_splash, exception_handler, qt_message_handler
 from utils.logging_config import setup_logger
-from utils.startup_dialogs import show_palette_config_dialog
+from utils.ui.startup_dialogs import show_palette_config_dialog
 from utils.message_manager import MessageManager
 from utils.status_manager import update_status_label
