+++ utils/ui/ui_setup.py
@@ -14,7 +14,7 @@ from ui_files.ui_main_window import Ui_Form
 from ui_files.visualization_3d import initialize_3d_view, clear_canvas, load_rob_files
 from utils import global_vars
 from utils.status_manager import update_status_label
-from utils.ui_helpers import (CustomDoubleValidator, handle_scanner_status,
+from utils.ui.ui_helpers import (CustomDoubleValidator, handle_scanner_status,
                              set_wordlist, open_page, Page, open_password_dialog, leave_settings_page,
                              open_file, save_open_file, execute_command, open_folder_dialog, 
                              open_file_dialog, set_settings_line_edits, check_key_or_password, clear_filters,
@@ -28,6 +28,8 @@ from utils.audio.audio import (spawn_play_stepback_warning_thread, kill_play_ste
 from utils.updater import check_for_updates
 from utils.app_control import restart_app, exit_app
 from utils.server.UR20_Server_functions import scanner_signals
+from utils.ui.notification_popup import check_zwischenlage_status
+from utils.ui.ui_helpers import check_palette_clearing_status
 
 # Add logger
 logger = logging.getLogger(__name__)
@@ -275,7 +277,7 @@ def start_background_tasks():
     warning_sound_thread.start()
     
     # Start zwischenlage popup monitor
-    from utils.notification_popup import check_zwischenlage_status
+    check_zwischenlage_status()
     
     # Create a timer to check zwischenlage status every 500ms
     zwischenlage_timer = QTimer(global_vars.main_window)
@@ -283,7 +285,6 @@ def start_background_tasks():
     zwischenlage_timer.start(500)  # Check every 500ms
     
     # Create a timer to check palette clearing status every 1000ms
-    from utils.ui_helpers import check_palette_clearing_status
     palette_clear_timer = QTimer(global_vars.main_window)
     palette_clear_timer.timeout.connect(check_palette_clearing_status)
     palette_clear_timer.start(1000)  # Check every 1000ms
